extract buildSearchRequest into dedicated file; place tests next to file
https://coveord.atlassian.net/browse/KIT-317
