+++ packages/headless/src/features/search/search-request-builder.test.ts
@@ -1,12 +1,12 @@
-import {createMockState} from '../../../test/mock-state';
-import {buildMockFacetRequest} from '../../../test/mock-facet-request';
-import {buildMockNumericFacetRequest} from '../../../test/mock-numeric-facet-request';
-import {buildMockDateFacetRequest} from '../../../test/mock-date-facet-request';
-import {buildMockCategoryFacetRequest} from '../../../test/mock-category-facet-request';
-import {buildMockFacetOptions} from '../../../test/mock-facet-options';
-import {buildMockFacetResponse} from '../../../test/mock-facet-response';
-import {SearchAppState} from '../../../state/search-app-state';
-import {buildSearchRequest} from '../../../features/search/search-actions';
+import {createMockState} from '../../test/mock-state';
+import {buildMockFacetRequest} from '../../test/mock-facet-request';
+import {buildMockNumericFacetRequest} from '../../test/mock-numeric-facet-request';
+import {buildMockDateFacetRequest} from '../../test/mock-date-facet-request';
+import {buildMockCategoryFacetRequest} from '../../test/mock-category-facet-request';
+import {buildMockFacetOptions} from '../../test/mock-facet-options';
+import {buildMockFacetResponse} from '../../test/mock-facet-response';
+import {SearchAppState} from '../../state/search-app-state';
+import {buildSearchRequest} from './search-request-builder';
 
 describe('search request', () => {
   let state: SearchAppState;
